NetEQ-related bug in ACM
Fixing a bug when creating new NetEQ slave instances in ACM.
The old code called WebRtcNetEQ_GetCurrentDelay() for the
master instance to get a delay value for WebRtcNetEQ_SetExtraDelay().
This is wrong, since WebRtcNetEQ_GetCurrentDelay() reports on the
current total buffer length, while WebRtcNetEQ_SetExtraDelay() is
the extra delay that is desired to in order to sync with video.

The fix includes keeping the extra delay value in a member variable
in the ACMNetEQ class.

Review URL: http://webrtc-codereview.appspot.com/295001

git-svn-id: http://webrtc.googlecode.com/svn/trunk/src@1001 4adac7df-926f-26a2-2b94-8c16560cd09d