Refactoring of TestAllCodecs
ACM testing consists of seven individual tests. Up til now we haven't used gtest everywhere, and many of the tests needs some rewriting to follow the style guide.

I've started with this tests, doing formatting, adding the test as a separate test which can now either succeed of fail in a proper way.

Still to do in this test is handling of input file, but that will be changed in a separate CL, because all tests uses the  PCMFile class that will be affected by the change.

BUG=none
TEST=audio_coding_module_test, ACM_AUTO_TEST and ACM_TEST_ALL_CODECS.

Review URL: https://webrtc-codereview.appspot.com/646011

git-svn-id: http://webrtc.googlecode.com/svn/trunk/src@2416 4adac7df-926f-26a2-2b94-8c16560cd09d