Refactoring a bit the switch between C++ OpenSL ES and Java one. Now in order to switch, we just need to flip the WEBRTC_ANDROID_OPENSLES in ./modules/audio_device/Android.mk and ./voice_engine/Android.mk
All the files are included in the build. Further in the future, if
the C++ OpenSL ES realization will be better, we might provide this
switch for runtime instead of compile time.

Although as much as I'd like to flip this flag at a single place,
putting it inside android-webrtc.mk where all the common flags are
set doesn't work as intended, cause it doesn't trigger the rebuild
of these two modules. So we leave it as it is.